+++ internal/server/server.go
@@ -616,7 +616,7 @@ func (s *Server) buildResolvers() []parser.Resolver {
 
 	r := make([]parser.Resolver, len(s.ComponentResolvers))
 	for i, cr := range s.ComponentResolvers {
-		r[i] = parser.Resolver{Match: cr.Match, Resolve: cr.Resolve, Prefix: cr.Prefix, NoFollow: cr.NoFollow, Anchored: cr.Anchored}
+		r[i] = parser.Resolver{Match: cr.Match, Resolve: cr.Resolve, Prefix: cr.Prefix, NoFollow: cr.NoFollow, Anchored: cr.Anchored, DynamicIfMissing: cr.DynamicIfMissing}
 	}
 
 	s.cachedResolverSet = parser.BuildResolverSet(r)